reorganize into sphinxcontrib.confluencebuilder namespace
Move the implementation of sphinxcontrib-confluencebuilder into a single
namespace: `sphinxcontrib.confluencebuilder`.
