--- sphinxcontrib/builders/__init__.py
@@ -1,7 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-    sphinxcontrib.builders
-    ~~~~~~~~~~~~~~~~~~~~~~
-
-    Custom docutils builders.
-"""
