--- sphinxcontrib/writers/__init__.py
@@ -1,7 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-    sphinxcontrib.writers
-    ~~~~~~~~~~~~~~~~~~~~~
-
-    Custom docutils writers.
-"""
